Dispatching tests according to method change
