+++ frictionless/resource/__spec__/test_validate_schema.py
