--- frictionless/validator/__spec__/resource/test_format.py
@@ -1,9 +1,0 @@
-from frictionless.resources import TableResource
-
-# General
-
-
-def test_resource_validate_format():
-    resource = TableResource(path="data/table.csv", format="csv")
-    report = resource.validate()
-    assert report.valid
